+++ resource/command.go
@@ -66,8 +66,8 @@ func (c *Command) GetExec() string {
 	return c.id
 }
 
-func (c *Command) Validate(sys *system.System) []TestResult {
-	ctx := context.WithValue(context.Background(), idKey{}, c.ID())
+func (c *Command) Validate(ctx context.Context, sys *system.System) []TestResult {
+	ctx = context.WithValue(ctx, idKey{}, c.ID())
 	skip := c.Skip
 
 	if c.Timeout == 0 {